Remove get/set of voicemail ringtone uri in shared preferences.
Prior to Android P, TelephonyManager#setVoicemailRingtoneUri was used
by the dialer app to set the voicemail notification sound played when
the platform got a new voicemail notification.  Likewise,
getVoicemailRingtoneUri was used to retrieve the set value.

Prior to P this was just saved in the shared prefs, but after P a
migration was done to move the shared preference to the
NotificationChannel#getSound for the voicemail notification.  If, however,
you called `setVoicemailRingtoneUri` it was still possible to change the
shared preference and have that migrated to be set on the notification
channel, causing a cross-profile exploit.

In the current world, the notifications for voicemail are NOT posted in
Telephony any more, and are instead associated with the notification
channel for voicemail IN the dialer app.  On the off chance a dialer does
not show the voicemail notification, Telephony can post it as well, but
at this point the related sound is expected to be associated with the
notification channel.

To mitigate this cross-profile vulnerability:
1. Ensure TelephonyManager#setVoicemailRingtoneUri does not save to shared
preferences any more.
2. Ensure the TelephonyManager#getVoicemailRingtoneUrigetRingtoneUri ONLY
queries from the notification channel, and not from the shared
preferences since that is not used.  This ensures we can never return a
bad URI set via the setter.
3. Remove the code in migrateVoicemailNotificationSettings which will take
the shared preference and migrate it over to the channel; this is not
needed as realistically ANY device from P would have updated LONG ago and
had its notification setting migrated to the channel anyways.

Test: Change the default voicemail notification channel sound on
"phone services"; verify that Dialer can still get this value.
Test: Changed the voicemail notification channel in the dialer app so that
it has a different value; verify that voicemail notifications use the
correct sound.
Flag: EXEMPT security patch.
Bug: 325030433
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:8e47af093625b997ffb8ca0379a4a56c02ddeb20)
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:5057ee3badc307777a961daafe94471ffcaf577e)
Merged-In: I7252c692eb2a5ff4b4fcbddba77425cb423539f3
(cherry picked from commit 5b19e54e8ad9d2c707ecc29ca78e3cc173dbb366)
Change-Id: I7252c692eb2a5ff4b4fcbddba77425cb423539f3